Add many symbols for the 74LVC1G57 and 74LVC1G58 multi-function gates.
These allow you to simply place a gate symbol, such as an AND gate, without
worrying about how the 1G57/1G58 chip has to be wired up.  The symbol will
automatically connect the correct pins for you - usually using a hidden pin
for the 3rd 1G57/1G58 input, which is automatically tied to one of the other
inputs or to VCC or GND.

There is a separate symbol for each possible configuration given in the
datasheet.  De Morgan variants are also included.

Also includes "multiplexer" symbols for the 74LVC1G97 and 74LVC1G98
multi-function gates.  De Morgan variants are also included, but not any of
the other configurations.