Fix LED:Neopixel_THT symbol.  The pins were numbered in reverse order.
According to the quoted "datasheet", https://www.adafruit.com/product/1938 ,
the labelled photo shows "Data In" is closest to the the flat of the LED.
According to the standard library footprint LED_THT:LED_D5.0mm-4_RGB
and variants, that is pin 1.

An alternative datasheet for a compatible part from a different vendor is:
https://cdn.sparkfun.com/datasheets/Components/LED/COM-12877.pdf
That clearly labels "Data In" as pin 1, and matches the Adafruit pinout.

However, the symbol had "Data In" as pin 4.  This commit changes it to
be pin 1.

The other pins were changed in the same way and for the same reasons.
Pins 1 and 4 swap over, and pins 2 and 3 swap over.