Add symbols for ideal voltage, current, and controlled sources
Although voltage and current symbols already exist in the SPICE
libraries, they are not typical representations. This is especially
the case for the controlled-type sources.

I didn't include SPICE functionality for the ideal voltage/current
symbols as these already exist in the Simulation_SPICE library.
Also, because the controlled-type sources are generic (either current-
controlled or voltage-controlled) I didn't add the associated SPICE primitive.

A couple of violations with the klc-checker popped up, but they shouldn't
be an issue. All symbols violated not having any footprint filters
(they shouldn't as they are generic symbols, no?) and the part name
was repeated in the description (because the part name was a letter, having
'V' as in 'voltage' or 'I' as in 'ideal' resulted in a violation).