Rework D-Subminiature symbols
The shell size letter (B in DB25) were wrong for all of those symbols.
What lot of people call "DB-9" is actually a "DE-9" connector.
The same applies to the other shell sizes present in the library.
While at it, also rename the symbols to Plug/Receptacle. But keep the
male/female terminology in the keywords so they can be found in the search
using those terms.

I did quite a bit of research on which name is correct for D-Sub type
connectors. There does not seem to be a consensus among manufacturers.
Würth, Molex, NorComp and Assmann exclusively use male/female, but others like
TE, Amphenol, Cannon, JAE, 3M use Pin/Socket or Plug/Receptacle to
identify their connectors.

All of those words have been added to the description.