Convert all symbols to v7 format
KiCad7 is only a few weeks from being release. We expect contributors to
start using that version soon. Since the v7 and v6 format are not 100%
compatible, saving just a single symbol with KiCad7 will result in a huge
diff that is hard to read.
For that reason, we just convert all files to v7 in one go.

From now on, the official kicad-symbols repository will only accept
contributions in the version 7 format.