CH340x: V3 should be power-output
On CH340x ICs, the V3 pin is the output of a 3.3V regulator,
which usually needs a decoupling cap to be inserted.

Change the pin type and swap the V3 and VCC pins around to
make it easier to connect the VCC pin to the VBUS from a USB
connector.

On the CH340X variant, move TNOW/nDTR output to the right side, in the
same place as the nDTR for the CH340C.

On the CH340E variant, place the TNOW output also in that position,
as this is an output that will probably go to the same place as the
serial data (or be NC).

Fixes: https://gitlab.com/kicad/libraries/kicad-symbols/-/issues/3188