Update format of all symbols to KiCad9
Even though KiCad9 can read older symbol files, this change is needed
to make future additions to the library easy to review.
As with every major version, the file format changes slighly. And so does
the ordering of elements. Without this bulk change a simple addition to the
library would create a huge diff. Those diffs are hard to read making the
life of the library team difficult.

The update was executed with:
find . -name "*.kicad_sym" -exec kicad-cli sym upgrade --force {} \;